[stable/kube-state-metrics] allow configuring collectors (#2124)
kube-state-metrics has a collector for each kind of Kubernetes resource
it collects metrics from. Each of those collectors directly influences
the RBAC requirements for the Pods. Additionally due to the nature of
how kube-state-metrics is versioned and released it is not easy to
determine which collectors are using "alpha" APIs, which are not
necessarily enabled in a cluster, but kube-state-metrics by default has
all collectors enabled.